Corrigindo Bug da Home Duplicada, Orientacao e Delete de Imagem

+++ lib/models/section.dart
@@ -75,7 +75,8 @@ class Section extends ChangeNotifier {
     }
 
     for(final original in originalItems){
-      if(!items.contains(original)){
+      if(!items.contains(original)
+          && (original.image as String).contains('firebase')){
         try {
           final ref = await storage.getReferenceFromUrl(
               original.image as String
@@ -96,13 +97,15 @@ class Section extends ChangeNotifier {
   Future<void> delete() async {
     await firestoreRef.delete();
     for(final item in items){
-      try {
-        final ref = await storage.getReferenceFromUrl(
-            item.image as String
-        );
-        await ref.delete();
-      // ignore: empty_catches
-      } catch (e){}
+      if((item.image as String).contains('firebase')){
+        try {
+          final ref = await storage.getReferenceFromUrl(
+              item.image as String
+          );
+          await ref.delete();
+          // ignore: empty_catches
+        } catch (e){}
+      }
     }
   }
